try/requirements.txt
google-cloud-storage==2.7.0
plotly==5.11.0
streamlit_lottie==0.0.3
